requirements.txt
pyyaml==3.13
mock==2.0.0